refactor:create UserApiService
